+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `uefi::proto::shim` is now available on 32-bit x86 targets.
 - `Parity` and `StopBits` are now a newtype-enums instead of Rust enums. Their
   members now have upper-case names.
+- `FileSystem::try_exists` now returns `FileSystemResult<bool>`.
 
 ## uefi-macros - [Unreleased]